Add intelligent book negotiation addition/removal
Now the switchboard intelligently trims back negotiations from its book
before sending it out. It does so in such a way that minimizes nodes
from having to wait for fruitless connections. For example, if the
switchboard sees someone made an answer to an offer, it'll remove that
offer, so others won't try to make duplicate answers that the original
offerer will never use.

In tests, this seems to have sped up connection pools by like 4x.