Perform a massive update
This reworks the whole way the switchboard deals with things. The
request and response shape has been changed. In real life, this change,
in addition to accompanying changes in network, results in a speedup
from initial time taken to connect being around 10 or 15 seconds, to
well within a second usually.

Since this change represents an API change as well as a near complete
rewrite, this will precede a release to 0.2.